backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12
google-generativeai==0.8.3
aiosmtplib==3.0.2
email-validator==2.2.0
python-dotenv==1.0.1
